Create switch_Case.java
You have been given the following information:				
Nairobi is the city under the sun. The only times it rains is during the months of April to May, and during the months of July to October. During the rainy months, the cost of living goes down by a factor of 5. In the dry season, the cost of living goes up by a factor of 10. 
Starting with 1 as the month of January, get current month and current cost of living from a user, and use a switch statement to determine the cost of living for that month.